requirements.txt
numpy==1.19.4
pdf2image==1.14.0
Pillow==8.0.1
pkg-resources==0.0.0
